# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/api/v1/basic/city": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础数据"
                ],
                "summary": "根据省代码获取城市",
                "parameters": [
                    {
                        "type": "string",
                        "description": "省代码",
                        "name": "province_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/basic/district": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础数据"
                ],
                "summary": "根据城市代码获取区",
                "parameters": [
                    {
                        "type": "string",
                        "description": "城市代码",
                        "name": "city_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/api/v1/basic/street": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础数据"
                ],
                "summary": "根据区代码获取街道",
                "parameters": [
                    {
                        "type": "string",
                        "description": "区代码",
                        "name": "district_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/common/verificationcode": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公共接口"
                ],
                "summary": "获取验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "手机号码",
                        "name": "tel",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.ResponseData": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "返回数据",
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                },
                "status": {
                    "description": "成功失败标志；true：成功 、false：失败",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}